enhancement/moved document-type.json to constants folder and corrected the shown blood type on patient profile
