--- src/pages/DashboardAdmin/components/constants/document-type.json
@@ -1,17 +1,0 @@
-[
-    {
-      "id": "Tipo de Documento",
-      "value": "Tipo de Documento",
-      "isDisabled": true
-    },
-    {
-      "id": "national_id",
-      "value": "Cédula",
-      "isDisabled": false
-    },
-    {
-      "id": "passport",
-      "value": "Pasaporte",
-      "isDisabled": false
-    }
-]
