+++ src/pages/RegisterPage/index.tsx
@@ -9,8 +9,8 @@ import HospiCloudLogo from "../../resources/HospiCloudLogo.svg";
 import { registerPatient } from "../../api/users/index";
 import { useEffect } from "react";
 import { useForm } from "react-hook-form";
-import DocumentType from "./document-type.json";
-import BloodyType from "./blood-type.json";
+import DocumentType from "../../constants/document-type.json";
+import BloodyType from "../../constants/blood-type.json";
 import { INewPatient } from "../../models/IPatient";
 import { useHistory } from "react-router";
 import routes from "../../router/constantRoutes.json";
